requirements.txt
pytest==7.4.3
pytest-xdist==3.5.0
pytest-html==4.1.1
requests==2.31.0
allure-pytest==2.13.2
pyyaml==6.0.1
loguru==0.7.2
pydantic==2.5.0
playwright==1.40.0
selenium==4.15.2
webdriver-manager==4.0.1
kafka-python==2.0.2
grpcio==1.59.3
grpcio-tools==1.59.3
